+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/common/find_previous_tag.rb
@@ -8,6 +8,7 @@ module Fastlane
     class FindPreviousTagAction < Action
       def self.run(params)
         tag_pattern = params[:pattern]
+        exclude_pattern = params[:exclude]
 
         # Make sure we have all the latest tags fetched locally
         Actions.sh('git', 'fetch', '--tags', '--force') { nil }
@@ -17,6 +18,7 @@ module Fastlane
         # Finally find the previous tag matching the provided pattern, and that is not the current commit
         git_cmd = %w[git describe --tags --abbrev=0]
         git_cmd += ['--match', tag_pattern] unless tag_pattern.nil?
+        git_cmd += ['--exclude', exclude_pattern] unless exclude_pattern.nil?
         git_cmd += ['--exclude', current_commit_tag] unless current_commit_tag.empty?
         Actions.sh(*git_cmd) { |exit_status, stdout, _| exit_status.success? ? stdout.chomp : nil }
       end
@@ -38,7 +40,8 @@ module Fastlane
           Uses `git describe --tags --abbrev=0 --match … --exclude …` to find the previous git tag
           reachable from the current commit and that matches a specific naming pattern
 
-          e.g. `find_previous_tag(pattern: '12.3.*.*')`, `find_previous_tag(pattern: '12.3-rc-*')`
+          e.g. `find_previous_tag(pattern: '12.3.*.*')`, `find_previous_tag(pattern: '12.3-rc-*')`,
+          `find_previous_tag(pattern: 'v*', exclude: '*beta*')`
         DETAILS
       end
 
@@ -46,6 +49,11 @@ module Fastlane
         [
           FastlaneCore::ConfigItem.new(key: :pattern,
                                        description: 'The _fnmatch_-style pattern to use when searching for the previous tag',
+                                       optional: true,
+                                       default_value: nil,
+                                       type: String),
+          FastlaneCore::ConfigItem.new(key: :exclude,
+                                       description: 'An _fnmatch_-style pattern of tags to exclude from the search (maps to `git describe --exclude`)',
                                        optional: true,
                                        default_value: nil,
                                        type: String),